refactor: abstract main useeffect asteroid logic
