+++ src/components/Asteroid/Fragments/Fragment-Asteroid-1.tsx
