+++ src/components/Asteroid/Fragments/Fragment-Asteroid-2.tsx
