+++ src/components/Asteroid/Fragments/Fragment-Asteroid-3.tsx
