+++ src/components/Asteroid/Fragments/Fragment-Asteroid-4.tsx
